requirements.txt
Django==5.2.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-filter==24.3
django-cors-headers==4.5.0
Pillow==10.4.0
python-decouple==3.8
mysqlclient==2.2.7
whitenoise==6.8.1
gunicorn==23.0.0